Some progress and additional refactors. It's probably about time to go all the way and make implicitly generated code more first-class.
