+++ compiler/cx-typechecker-data/src/format.rs
@@ -2,6 +2,96 @@ use super::ast::*;
 use crate::cx_types::{CXFunctionPrototype, CXParameter, CXType, CXTypeKind};
 use std::fmt::{Display, Formatter, Result};
 
+pub(crate) fn type_mangle(ty: &CXType) -> String {
+    let mut mangled = String::new();
+
+    match &ty.kind {
+        CXTypeKind::PointerTo { inner_type, .. } => {
+            mangled.push('P');
+            mangled.push_str(&type_mangle(inner_type));
+        }
+        CXTypeKind::StrongPointer {
+            inner_type,
+            is_array,
+        } => {
+            mangled.push('S');
+            if *is_array {
+                mangled.push('A');
+            } else {
+                mangled.push('P');
+            }
+            mangled.push_str(&type_mangle(inner_type));
+        }
+        CXTypeKind::MemoryReference(inner_type) => {
+            mangled.push('R');
+            mangled.push_str(&type_mangle(inner_type));
+        }
+        CXTypeKind::Opaque { size, .. } => {
+            mangled.push('O');
+            mangled.push_str(&size.to_string());
+        }
+        CXTypeKind::Array { size, inner_type } => {
+            mangled.push('A');
+            mangled.push_str(&size.to_string());
+            mangled.push('_');
+            mangled.push_str(&type_mangle(inner_type));
+        }
+        CXTypeKind::VariableLengthArray { _type, .. } => {
+            mangled.push_str("V_a");
+            mangled.push_str(&type_mangle(_type));
+        }
+        CXTypeKind::Function { prototype } => {
+            mangled.push('F');
+            mangled.push_str(&type_mangle(&prototype.return_type));
+            for param in &prototype.params {
+                mangled.push_str(&type_mangle(&param._type));
+            }
+            mangled.push(prototype.var_args as u8 as char);
+        }
+        CXTypeKind::Structured { fields, .. } => {
+            mangled.push('S');
+            mangled.push_str(&fields.len().to_string());
+            mangled.push('_');
+            for field in fields {
+                mangled.push_str(&type_mangle(&field.1));
+            }
+        }
+        CXTypeKind::Union { variants, .. } => {
+            mangled.push('U');
+            mangled.push_str(&variants.len().to_string());
+            mangled.push('_');
+            for variant in variants {
+                mangled.push_str(&type_mangle(&variant.1));
+            }
+        }
+        CXTypeKind::TaggedUnion { variants, .. } => {
+            mangled.push('T');
+            mangled.push_str(&variants.len().to_string());
+            mangled.push('_');
+            for variant in variants {
+                mangled.push_str(&type_mangle(&variant.1));
+            }
+        }
+        CXTypeKind::Integer { bytes, signed } => {
+            mangled.push('I');
+            mangled.push_str(&bytes.to_string());
+            mangled.push(if *signed { 's' } else { 'u' });
+        }
+        CXTypeKind::Float { bytes } => {
+            mangled.push('F');
+            mangled.push_str(&bytes.to_string());
+        }
+        CXTypeKind::Bool => {
+            mangled.push('B');
+        }
+        CXTypeKind::Unit => {
+            mangled.push('v');
+        }
+    }
+
+    return mangled;
+}
+
 // Helper struct for indented formatting of TCExpr
 struct TCExprFormatter<'a> {
     expr: &'a TCExpr,
