+++ compiler/cx-typechecker/src/expr_checking/binary_ops.rs
@@ -1,8 +1,8 @@
-use crate::casting::{add_coercion, coerce_value, implicit_cast};
 use crate::environment::TCEnvironment;
+use crate::expr_checking::casting::{add_coercion, coerce_value, implicit_cast};
+use crate::expr_checking::typechecker::typecheck_expr;
 use crate::log_typecheck_error;
-use crate::type_mapping::contextualize_template_args;
-use crate::typechecker::typecheck_expr;
+use crate::type_completion::type_mapping::contextualize_template_args;
 use cx_parsing_data::parse::ast::{CXBinOp, CXCastType, CXExpr, CXExprKind};
 use cx_typechecker_data::ast::{TCExpr, TCExprKind};
 use cx_typechecker_data::cx_types::{CXType, CXTypeKind, same_type};
