+++ compiler/cx-typechecker/src/type_completion/mod.rs
@@ -1,0 +1,3 @@
+pub mod precontextualizing;
+pub mod templates;
+pub mod type_mapping;
